+++ src/sklearndf/transformation/extra/wrapper/__init__.py
@@ -1,0 +1,6 @@
+"""
+DF wrapper classes for additional 3rd party transformers that implement the
+`scikit-learn` interface.
+"""
+
+from ._wrapper import *
